% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_plotsvm.R
\name{gg_plotsvm}
\alias{gg_plotsvm}
\title{gg_plotsvm}
\usage{
gg_plotsvm(model, data, x, y, ...)
}
\arguments{
\item{model}{kernlab model}

\item{data}{data frame used in kernlab model}

\item{x}{feature 1 from data in model}

\item{y}{feature 2 from data in model}

\item{...}{parameters passed to aes() such as color}
}
\value{
a ggplot object
}
\description{
Plot an SVM model created by \code{kernlab}
}
\examples{
\dontrun{
data(iris)
iris_data <- iris[iris$Species != "setosa", ]
iris_data$Species <- factor(iris_data$Species)
df = iris_data[,c(1,2,5)]
require(kernlab)
# scaled=TRUE is default for ksvm
model2 <- kernlab::ksvm(Species ~ ., df, kernel = "vanilladot", C = 1)
gg_plotsvm(model2, df,
              Sepal.Length, Sepal.Width, color=Species)
model3 <- kernlab::ksvm(Species ~ ., df, kernel = "vanilladot", scaled=FALSE, C = 1)
gg_plotsvm(model3, df,
              Sepal.Length, Sepal.Width, color=Species)
}
}
