% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_plotsvm.R
\name{gg_plotsvm_contour}
\alias{gg_plotsvm_contour}
\title{gg_plotsvm_contour}
\usage{
gg_plotsvm_contour(grid, bins = 10, legend = FALSE)
}
\arguments{
\item{grid}{a grid created by svmgrid}

\item{bins}{number of bins for contouring}

\item{legend}{logical, default FALSE, display legend}
}
\value{
ggplot object
}
\description{
Adds contours to an svm ggplot object using a grid
}
\details{
\itemize{
\item You can change the contour colors by adding a ggplot call to
\code{scale_fill_viridis_d(option="<one of the viridis options>")}
}
}
\examples{
\dontrun{
iris_data <- iris[iris$Species != "setosa", ]
iris_data$Species <- factor(iris_data$Species)
df = iris_data[,c(1,2,5)]
require(kernlab)
require(tidyverse)
# scaled=TRUE is default for ksvm
model2 <- kernlab::ksvm(Species ~ ., df, kernel = "vanilladot", C = 1)
gg_plotsvm(model2, df,
              Sepal.Length, Sepal.Width, color=Species)
model3 <- kernlab::ksvm(Species ~ ., df, kernel = "vanilladot", scaled=FALSE, C = 1)
gg_plotsvm(model3, df,
              Sepal.Length, Sepal.Width, color=Species) +
   gg_plotsvm_contour(grid = svmgrid(df, model) )
}
}
