% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ginit.R
\name{ginisearch}
\alias{ginisearch}
\title{ginisearch}
\usage{
ginisearch(.data, .var, .target)
}
\arguments{
\item{.data}{a data frame or something that can be converted to a data frame}

\item{.var}{variable to be searched on}

\item{.target}{categorical variable in dataset}
}
\value{
tibble with the cutpoints of search variable and their gini indices
}
\description{
ginisearch
}
\examples{
\dontrun{
require(tidyverse)
z2 = Table_8_1 \%>\% mutate(x2=(GlucoseLevel<159))
z2
q2 = addmargins(table(z2$x2, z2$Diabetes))
q2
ginisearch(q2)
}
}
