% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ginit.R
\name{ginit}
\alias{ginit}
\title{ginit}
\usage{
ginit(t)
}
\arguments{
\item{t}{table of variables for gini Index}
}
\value{
numeric Gini index.
}
\description{
Gini index from a table with margins added
}
\examples{
\dontrun{
require(tidyverse)
z2 = Table_8_1 \%>\% mutate(x2=(GlucoseLevel<159))
z2
q2 = table(z2$x2, z2$Diabetes) \%>\% addmargins
q2
ginit(q2)
}

}
