\name{HEADS-package}
\alias{HEADS-package}
\alias{HEADS}
\docType{package}
\title{HEADS}
\description{
Data files for Healthcare Engineering Analytics and Decision Sciences
for use in the exercises.
}
\details{
Data files referenced in the text have been created as R data files
accessible with the \code{data()} function. R notebooks detail use of these files.
}
\author{
\packageAuthor{HEADS}

Maintainer: \packageMaintainer{HEADS}
}
\references{
Mehrotra, Sanjay, Kevin Bui, and Hari Balasubramanian (2024). Healthcare Engineering,
Analytics, and Decision Sciences: An Introduction. December 11, 2024.
\href{https://sites.google.com/view/healthcareengineering/}{Healthcare Engineering}
}
\keyword{package}

\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{HEADS}}

\section{Documentation Index}{\packageIndices{HEADS}}

\seealso{
\href{https://connect.informs.org/discussion/new-textbook-on-healthcare-engineering-resending}{INFORMS New Textbook Notice}
}
\examples{
help(HEADS)
}
