% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ginit.R
\name{plot_ginisearch}
\alias{plot_ginisearch}
\title{plot_ginisearch}
\usage{
plot_ginisearch(.data)
}
\arguments{
\item{.data}{a tibble returned by ginisearch}
}
\value{
a ggplot object
}
\description{
plot_ginisearch
}
\examples{
\dontrun{
z2 = Table_8_1
z2$x2= (GlucoseLevel<159))
q2 = table(z2$x2, z2$Diabetes) \%>\% addmargins
ginit(q2)
}

}
