% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_plotsvm.R
\name{svmgrid}
\alias{svmgrid}
\title{svmgrid}
\usage{
svmgrid(.data, model, ng = 100)
}
\arguments{
\item{.data}{a data frame with two columns for gridding}

\item{model}{a model object produced by ksvm}

\item{ng}{number of points in each direction of grid}
}
\value{
data frame with 3 columns;
\itemize{
\item the grid points x and y for the columns of .data,
\item decision values predicted using the ksvm model
}
}
\description{
svmgrid
}
\examples{
\dontrun{
data(iris)
iris_data <- iris[iris$Species != "setosa", ]
iris_data$Species <- factor(iris_data$Species)
df = iris_data[,c(1,2,5)]
require(kernlab)
model <- kernlab::ksvm(Species ~ ., data = df, kernel = "vanilladot", scaled=FALSE, C = 1)
grid = svmgrid(df, model)
}
}
