% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_plotsvm.R
\name{svmmargin}
\alias{svmmargin}
\title{svmmargin}
\usage{
svmmargin(model)
}
\arguments{
\item{model}{a kernlab::svm model; only an unscaled model can be relied on}
}
\value{
margin, invisible
}
\description{
margin of a kernlab::svm model with linear kernel (\code{vanilladot});
}
\details{
margin is the distance of the best separating hyperplane to the closest
points of the dataset. The margin should be the same on either side. The closest points should be
symmetrically placed with respect to the hyperplane.

Margin is the distance of the best separating hyperplane to the closest
points of the dataset. The margin should be the same on either side.
The closest points should be symmetrically placed with respect to the hyperplane.
}
\examples{
\dontrun{
}
}
