% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_plotsvm.R
\name{svmunscale}
\alias{svmunscale}
\title{svmunscale}
\usage{
svmunscale(.data, model)
}
\arguments{
\item{.data}{object with columns to unscale}

\item{model}{keanlab svm model}
}
\value{
object with unscaled columns
}
\description{
svmunscale
}
\examples{
\dontrun{
}
}
