\name{Table_5_1}
\alias{Table_5_1}
\docType{data}
\title{
Systolic BP Data for 33 Subjects from Table 5.1
}
\description{
Age and Systolic Blood Pressure measurements for 33 subjects.
}
\usage{data("Table_5_1")}
\format{
  A data frame with 33 observations on the following 2 variables.
  \describe{
    \item{\code{Age}}{a numeric vector, integer age of subject}
    \item{\code{SBP}}{a numeric vector, systolic blood pressure measurement}
  }
}
\details{
33 adult women not treated with any medication.
}
\source{
Mehrotra, Sanjay, Kevin Bui, and Hari Balasubramanian (2024). Healthcare Engineering,
Analytics, and Decision Sciences: An Introduction. December 11, 2024.
(https://sites.google.com/view/healthcareengineering/), page 133.
}
\references{
Mehrotra, Sanjay, Kevin Bui, and Hari Balasubramanian (2024). Healthcare Engineering,
Analytics, and Decision Sciences: An Introduction. December 11, 2024.
(https://sites.google.com/view/healthcareengineering/)
}
\examples{
data(Table_5_1)
}
\keyword{datasets}
