\name{Table_5_10}
\alias{Table_5_10}
\docType{data}
\title{
Relation between Height, Weight, and Blood Fat Content
}
\description{
Data was taken by Burkhardt determine the relationship between blood fat content and a patient’s age and weight,
displayed in the table 5.10.}
\usage{data("Table_5_10")}
\format{
  A data frame with 25 observations on the following 3 variables.
  \describe{
    \item{\code{Weight}}{a numeric vector, weight (kg)}
    \item{\code{Age}}{a numeric vector, age (years)}
    \item{\code{BloodFatContent}}{a numeric vector, blood fat content; no units specified}
  }
}
\details{
Table 5.10
}
\source{
Mehrotra, Sanjay, Kevin Bui, and Hari Balasubramanian (2024). Healthcare Engineering,
Analytics, and Decision Sciences: An Introduction. December 11, 2024.
(https://sites.google.com/view/healthcareengineering/), Chapter 5.
}
\references{
Dataset source: (http://people.sc.fsu.edu/jburkardt/datasets/regression/regression.html)
}
\examples{
data(Table_5_10)
}
\keyword{datasets}
