\name{Table_5_11}
\alias{Table_5_11}
\docType{data}
\title{
Cholesterol and Diastolic Blood Pressure
}
\description{
Relation between Cholesterol Level and
Blood Pressure}
\usage{data("Table_5_11")
}
\format{
  A data frame with 19 observations on the following 2 variables.
  \describe{
    \item{\code{Cholesterol}}{a numeric vector}
    \item{\code{DiastolicBloodPressure}}{a numeric vector}
  }
}
\details{
Table 5.11
}
\source{
Mehrotra, Sanjay, Kevin Bui, and Hari Balasubramanian (2024). Healthcare Engineering,
Analytics, and Decision Sciences: An Introduction. December 11, 2024.
(https://sites.google.com/view/healthcareengineering/), Chapter 5.
}
\references{
Dataset source: (http://people.sc.fsu.edu/jburkardt/datasets/regression/regression.html)
}
\examples{
data(Table_5_11)
}
\keyword{datasets}
