\name{Table_5_12}
\alias{Table_5_12}
\docType{data}
\title{
Age and Blood Pressure
}
\description{
Relationship between age and systolic blood pressure according to the sample data
}
\usage{data("Table_5_12")}
\format{
  A data frame with 20 observations on the following 2 variables.
  \describe{
    \item{\code{Age}}{a numeric vector, Age (years)}
    \item{\code{SystolicBloodPressure}}{a numeric vector, (mmHg)}
  }
}
\details{
Table 5.12
}
\source{
Mehrotra, Sanjay, Kevin Bui, and Hari Balasubramanian (2024). Healthcare Engineering,
Analytics, and Decision Sciences: An Introduction. December 11, 2024.
(https://sites.google.com/view/healthcareengineering/), Chapter 5.
}
\references{
Dataset source: (http://people.sc.fsu.edu/jburkardt/datasets/regression/regression.html)
}
\examples{
data(Table_5_12)
}
\keyword{datasets}
