\name{Table_5_13}
\alias{Table_5_13}
\docType{data}
\title{
BMI and Systolic Blood Pressure
}
\description{
Relationship between body mass index (BMI) and systolic blood pressure according to the sample data
}
\usage{data("Table_5_13")}
\format{
  A data frame with 28 observations on the following 2 variables.
  \describe{
    \item{\code{BMI}}{a numeric vector, body mass index (BMI)}
    \item{\code{SystolicBP}}{a numeric vector, Systolic Blood Pressure (mmHg)}
  }
}
\details{
Table 5.13
}
\source{
Mehrotra, Sanjay, Kevin Bui, and Hari Balasubramanian (2024). Healthcare Engineering,
Analytics, and Decision Sciences: An Introduction. December 11, 2024.
(https://sites.google.com/view/healthcareengineering/), Chapter 5.
}
%\references{
%Dataset source: (http://people.sc.fsu.edu/jburkardt/datasets/regression/regression.html)
%}
\examples{
data(Table_5_13)
}
\keyword{datasets}
