\name{Table_5_14}
\alias{Table_5_14}
\docType{data}
\title{
Patient Length of Stay, Previous Admissions, and Satisfaction}
\description{
You want to find out whether or not length of stay and the number of previous admissions affect the
individual patient’s satisfaction score.
}
\usage{data("Table_5_14")}
\format{
  A data frame with 15 observations on the following 3 variables.
  \describe{
    \item{\code{LengthofStay}}{a numeric vector, length ofhospital stay (days)}
    \item{\code{PreviousAdmissions}}{a numeric vector, integewr number of previous admissions}
    \item{\code{PatientSatisfaction}}{a numeric vector, satisfaction score; average satisfaction=100}
  }
}
\details{
Table 5.14
}
\source{
Mehrotra, Sanjay, Kevin Bui, and Hari Balasubramanian (2024). Healthcare Engineering,
Analytics, and Decision Sciences: An Introduction. December 11, 2024.
(https://sites.google.com/view/healthcareengineering/), Chapter 5.
}
%\references{
%Dataset source: (http://people.sc.fsu.edu/jburkardt/datasets/regression/regression.html)
%}
\examples{
data(Table_5_14)
}
\keyword{datasets}
