\name{Table_5_5}
\alias{Table_5_5}
\docType{data}
\title{
Kidney transplant recipients, diagnoses, and waiting time
}
\description{
Kidney transplant recipients, their diagnoses, and waiting time data.
GM: glomerular disease; D: diabetes; HTN: hypertensive nephrosclerosis
}
\usage{data("Table_5_5")}
\format{
  A data frame with 30 observations on the following 2 variables.
  \describe{
    \item{\code{Diagnosis}}{a character vector with 3 possible values.}
    \item{\code{WaitingTime}}{a numeric vector, waiting time (days)}
  }
}
\details{
Table 5.5 in the source. Kidney transplant recipients, their diagnoses, and waiting time data.
GM: glomerular disease; D: diabetes; HTN: hypertensive nephrosclerosis.
}
\source{
Mehrotra, Sanjay, Kevin Bui, and Hari Balasubramanian (2024). Healthcare Engineering,
Analytics, and Decision Sciences: An Introduction. December 11, 2024.
(https://sites.google.com/view/healthcareengineering/), Chapter 5.
}
% \references{
%%  ~~ possibly secondary sources and usages ~~
% }
\examples{
data(Table_5_5)
}
\keyword{datasets}
