\name{Table_5_9}
\alias{Table_5_9}
\docType{data}
\title{
Catheter Length
}
\description{
For several young patients, catheters were fed from a principal vein into the heart. The necessary
catheter length was measured. The height and weight of the patients was noted.
}
\usage{data("Table_5_9")}
\format{
  A data frame with 12 observations on the following 3 variables.
  \describe{
    \item{\code{Height}}{a numeric vector, Height (in)}
    \item{\code{Weight}}{a numeric vector, Weight (lbs)}
    \item{\code{CatheterLength}}{a numeric vector, Catheter length (cm)}
  }
}
\details{
Table 5.9.
}
\source{
Mehrotra, Sanjay, Kevin Bui, and Hari Balasubramanian (2024). Healthcare Engineering,
Analytics, and Decision Sciences: An Introduction. December 11, 2024.
(https://sites.google.com/view/healthcareengineering/), Chapter 5.
}
\references{
Dataset source: (http://people.sc.fsu.edu/jburkardt/datasets/regression/regression.html)
}
\examples{
data(Table_5_9)
}
\keyword{datasets}
