\name{Table_8_2}
\alias{Table_8_2}
\docType{data}
\title{
Diabetes Patient Measurements (10)
}
\description{
A dataset of 10 patients with their glucose levels, BMIs, and diabetes statuses.}
\usage{data("Table_8_2")}
\format{
  A data frame with 10 observations on the following 3 variables.
  \describe{
    \item{\code{GlucoseLevel}}{a numeric vector, glucose level (mg/dL)}
    \item{\code{BMI}}{a numeric vector, Body mass index}
    \item{\code{Diabetes}}{a character vector, "negative" or "positive" for diabetes}
  }
}
\details{
Table 8.2
}
\source{
Mehrotra, Sanjay, Kevin Bui, and Hari Balasubramanian (2024). Healthcare Engineering,
Analytics, and Decision Sciences: An Introduction. December 11, 2024.
(https://sites.google.com/view/healthcareengineering/), Chapter 8, page 335
}
%\references{
%%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(Table_8_2)
}
\keyword{datasets}
